ai-insights-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
langchain==0.3.7
langchain-ollama==0.2.0
langchain-community==0.3.7
httpx==0.27.2
python-jose==3.3.0
python-dotenv==1.0.1